+++ tests/test_encoders.py
@@ -203,26 +203,42 @@ def test_xresnet_encoder(encoder, encoder_params):
 @pytest.mark.parametrize(
     ["encoder", "encoder_params"],
     [
-        [timm.B0Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B1Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B2Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B3Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B4Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B5Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B6Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.B7Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.MixNetXLEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.SKResNet18Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.SKResNeXt50Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.SWSLResNeXt101Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
-        [timm.HRNetW18Encoder, {"pretrained": False}],
-        [timm.HRNetW32Encoder, {"pretrained": False}],
-        [timm.HRNetW48Encoder, {"pretrained": False}],
-        [timm.DPN68Encoder, {"pretrained": False}],
-        [timm.DPN68BEncoder, {"pretrained": False}],
-        [timm.DPN92Encoder, {"pretrained": False}],
-        [timm.DPN107Encoder, {"pretrained": False}],
-        [timm.DPN131Encoder, {"pretrained": False}],
+        # [timm.B0Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B1Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B2Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B3Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B4Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B5Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B6Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.B7Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.MixNetXLEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.SKResNet18Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.SKResNeXt50Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.SWSLResNeXt101Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        # [timm.HRNetW18Encoder, {"pretrained": False}],
+        # [timm.HRNetW32Encoder, {"pretrained": False}],
+        # [timm.HRNetW48Encoder, {"pretrained": False}],
+        # [timm.DPN68Encoder, {"pretrained": False}],
+        # [timm.DPN68BEncoder, {"pretrained": False}],
+        # [timm.DPN92Encoder, {"pretrained": False}],
+        # [timm.DPN107Encoder, {"pretrained": False}],
+        # [timm.DPN131Encoder, {"pretrained": False}],
+        [timm.NFNetF0Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF1Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF2Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF3Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF4Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF5Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF6Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF7Encoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF0SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF1SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF2SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF3SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF4SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF5SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF6SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
+        [timm.NFNetF7SEncoder, {"pretrained": False, "layers": [0, 1, 2, 3, 4]}],
     ],
 )
 @torch.no_grad()
@@ -238,9 +254,10 @@ def test_timm_encoders(encoder, encoder_params):
     output = net(x)
     assert len(output) == len(net.channels)
     for feature_map, expected_stride, expected_channels in zip(output, net.strides, net.channels):
+        assert feature_map.size(0) == x.size(0)
         assert feature_map.size(1) == expected_channels
-        assert feature_map.size(2) * expected_stride == 256
-        assert feature_map.size(3) * expected_stride == 256
+        assert feature_map.size(2) * expected_stride == x.size(2)
+        assert feature_map.size(3) * expected_stride == x.size(3)
 
 
 @pytest.mark.parametrize(
